Module discord_sdk::lobby::search

Structs

SearchFilter

SearchQuery

A query used to search for lobbies that match a set of criteria.

SearchSort

Enums

LobbySearchCast

Determines how the search value is cast before comparison

LobbySearchComparison

The logical comparison to use when comparing the value of the filter key in the lobby metadata against the value provided to compare it against

LobbySearchDistance

The search distance from the current user’s region, the LobbySearchDistance::Default is to search in the current user’s region and adjacent regions.

SearchKey

SearchValue